--- execution/ft_memcmp.c
@@ -1,25 +1,0 @@
-/* ************************************************************************** */
-/*                                                                            */
-/*                                                        :::      ::::::::   */
-/*   ft_memcmp.c                                        :+:      :+:    :+:   */
-/*                                                    +:+ +:+         +:+     */
-/*   By: imittous <imittous@student.42.fr>          +#+  +:+       +#+        */
-/*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2021/11/11 16:14:52 by imittous          #+#    #+#             */
-/*   Updated: 2022/11/04 02:00:44 by imittous         ###   ########.fr       */
-/*                                                                            */
-/* ************************************************************************** */
-
-#include "mini_shell.h"
-
-int	ft_memcmp(char *s1, char *s2, size_t n)
-{
-	size_t	i;
-
-	if (n == 0)
-		return (0);
-	i = 0;
-	while (((unsigned char *)s1)[i] == ((unsigned char *)s2)[i] && i < (n - 1))
-		i++;
-	return (((unsigned char *)s1)[i] - ((unsigned char *)s2)[i]);
-}
