--- execution/ft_memcpy.c
@@ -1,28 +1,0 @@
-/* ************************************************************************** */
-/*                                                                            */
-/*                                                        :::      ::::::::   */
-/*   ft_memcpy.c                                        :+:      :+:    :+:   */
-/*                                                    +:+ +:+         +:+     */
-/*   By: imittous <imittous@student.42.fr>          +#+  +:+       +#+        */
-/*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2021/11/11 16:10:33 by imittous          #+#    #+#             */
-/*   Updated: 2022/11/04 01:53:47 by imittous         ###   ########.fr       */
-/*                                                                            */
-/* ************************************************************************** */
-
-#include "mini_shell.h"
-
-void	*ft_memcpy(void *dst, const void *src, size_t n)
-{
-	size_t		i;
-
-	i = 0;
-	if (!(char *)dst && !(char *)src)
-		return (0);
-	while (i < n)
-	{
-		((char *)dst)[i] = ((char *)src)[i];
-		i++;
-	}
-	return (dst);
-}
