--- execution/ft_strlen.c
@@ -1,23 +1,0 @@
-/* ************************************************************************** */
-/*                                                                            */
-/*                                                        :::      ::::::::   */
-/*   ft_strlen.c                                        :+:      :+:    :+:   */
-/*                                                    +:+ +:+         +:+     */
-/*   By: imittous <imittous@student.42.fr>          +#+  +:+       +#+        */
-/*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2021/11/01 18:52:58 by imittous          #+#    #+#             */
-/*   Updated: 2022/11/04 02:32:04 by imittous         ###   ########.fr       */
-/*                                                                            */
-/* ************************************************************************** */
-
-#include "mini_shell.h"
-
-size_t	ft_strlen(const char *str)
-{
-	int	i;
-
-	i = 0;
-	while (str[i])
-		i++;
-	return (i);
-}
