--- execution/ft_substr.c
@@ -1,38 +1,0 @@
-/* ************************************************************************** */
-/*                                                                            */
-/*                                                        :::      ::::::::   */
-/*   ft_substr.c                                        :+:      :+:    :+:   */
-/*                                                    +:+ +:+         +:+     */
-/*   By: imittous <imittous@student.42.fr>          +#+  +:+       +#+        */
-/*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2021/11/28 20:05:18 by imittous          #+#    #+#             */
-/*   Updated: 2022/11/04 02:32:15 by imittous         ###   ########.fr       */
-/*                                                                            */
-/* ************************************************************************** */
-
-#include "mini_shell.h"
-
-char	*ft_substr(char const *s, unsigned int start, size_t len)
-{
-	char	*p;
-	size_t	i;
-
-	if (!s)
-		return (NULL);
-	if (len > ft_strlen(&s[start]))
-		p = malloc((ft_strlen(&s[start]) + 1) * sizeof(char));
-	else
-		p = malloc((len + 1) * sizeof(char));
-	if (!p)
-		return (NULL);
-	if (start >= ft_strlen(s))
-		start = ft_strlen(s);
-	i = 0;
-	while (i < len && s[start + i])
-	{
-		p[i] = s[start + i];
-		i++;
-	}
-	p[i] = '\0';
-	return (p);
-}
