--- execution/mini_shell.h
@@ -1,20 +1,0 @@
-#ifndef MINI_SHELL_H
-#define MINI_SHELL_H
-
-#include <stdio.h>
-#include <stdlib.h>
-#include <string.h>
-#include <unistd.h>
-#include <sys/types.h>
-#include <sys/wait.h>
-#include <sys/stat.h>
-#include <fcntl.h>
-#include <readline/readline.h>
-#include <readline/history.h>
-
-void	*ft_memcpy(void *dst, const void *src, size_t n);
-int	ft_memcmp(char *s1, char *s2, size_t n);
-char	**ft_split(char const *s, char c);
-char	*ft_substr(char const *s, unsigned int start, size_t len);
-size_t	ft_strlen(const char *str);
-#endif
