--- output.txt
@@ -1,2 +1,0 @@
-This will be written to output.txt
-This will be written to standard output
