--- test.c
@@ -1,1 +1,0 @@
-#include <stdio.h>
